requirements.txt
django
djangorestframework
requests
psycopg2-binary
gunicorn
dj-database-url
whitenoise
beautifulsoup4
